+++ profile/profile.go
@@ -3,14 +3,14 @@ package profile
 import "github.com/codegram01/wingram-one/database"
 
 type Profile struct {
-	Id        int    `json:"id"`
+	Id        int64  `json:"id"`
 	Name      string `json:"name"`
-	AccountId int    `json:"account_id"`
+	AccountId int64  `json:"account_id"`
 }
 
 type ProfileReq struct {
 	Name      string `json:"name"`
-	AccountId int    `json:"account_id"`
+	AccountId int64  `json:"account_id"`
 }
 
 type Resource struct {
